docs: Clean up obsolete session reports and consolidate test documentation
Remove 52 obsolete session-specific summary and completion reports.
Consolidate redundant test documentation into comprehensive index files.

Deleted session reports:
- Root: 6 completion/implementation reports
- Tests/Playwright: 24 session summaries and fix reports
- docs/: 4 status/performance reports
- Test subdirs: 4 implementation summaries

Consolidated test documentation:
- Accessibility: 4 files -> A11Y_TESTS_INDEX.md
- Forms: 2 files -> FORM_TESTS_INDEX.md
- Media: 5 files -> MEDIA_TESTS_INDEX.md

All unique information preserved in consolidated files.
Historical context preserved in git history and CHANGELOG.md.

+++ Services/Sentinel/TestPhishingURLAnalyzer.cpp
@@ -169,7 +169,7 @@ static void test_suspicious_tlds()
             detected++;
     }
 
-    if (detected == suspicious_tld_urls.size())
+    if (static_cast<size_t>(detected) == suspicious_tld_urls.size())
         log_pass("Suspicious TLD detection");
     else
         log_fail("Suspicious TLDs", "Did not detect all suspicious TLDs");
